factor out transcript message cells into {Human,Assistant}MessageCell
Also remove CSS that was not necessary for correct text wrapping.
