+++ vscode/webviews/chat/cells/messageCell/BaseMessageCell.module.css
@@ -1,0 +1,26 @@
+.cell-container {
+    overflow: hidden;
+}
+
+/* Make footer take up less vertical space */
+.footer {
+    margin-top: calc(0.25*var(--cell-spacing));
+}
+
+/* Footer on hover */
+.cell-container .footer {
+    opacity: 0.25;
+}
+.cell-container:hover .footer {
+    opacity: 1;
+}
+
+/* Editing state */
+.disabled .footer {
+    opacity: 0;
+}
+.focused {
+    box-shadow: 0 -1px 0 var(--vscode-sideBarSectionHeader-border),
+                0 1px 0 var(--vscode-sideBarSectionHeader-border),
+                0 0 2rem var(--vscode-widget-shadow);
+}
