+++ vscode/webviews/chat/cells/messageCell/human/HumanMessageCell.module.css
@@ -1,0 +1,21 @@
+.content {
+    display: flex;
+    align-items: flex-start;
+    gap: var(--cell-spacing);
+    overflow: hidden;
+}
+.content--message {
+    flex: 1;
+    overflow: hidden;
+}
+
+[role="row"] .edit-button {
+    opacity: 0.25;
+}
+[role="row"]:hover .edit-button {
+    opacity: 1;
+}
+[role="row"]:is([aria-current="true"], [aria-disabled="true"]) .edit-button {
+    opacity: 0;
+    pointer-events: none;
+}
